move api converter to a dedicated module
